Add failing test for combining characters
In Unicode, a combining character is a character which can be stacked
on top of the character preceding it. For example:

- The character LATIN SMALL LETTER U (`u`) has the codepoint
  U+0075 assigned.
- The character COMBINING DIAERESIS, which looks similar to the
  symbol `¨` but is actually a combining character, has the
  codepoint U+0308 assigned.
- Writing both characters in sequence yields the letter `ü`,
  which looks just like `ü` but is actually *two* characters.

On a Mac, such combinations are very common, especially in filenames
due to an oddity in the HFS+ filesystem.

In the `applescript-json` library however, the `encodeString`
function always fails when the input contains a combining character.
In detail, the code assumes that inside the `repeat with ch` loop,
`ch` will be always one single character, and its `id` property will
always return an integer. However, in reality `ch` will contain more
than one character if a combining character is involved. Because of
that, the `id` property will return a list instead of an integer. The
code is not prepared to handle the list, which triggers the error.

This commit adds a simple test case for the “u followed by ̈” scenario
described above. It also includes the expected JSON output, which
would be `u\u0308`.